Fix dependency structure for ParameterOption
